refactor: 게임 상태를 LottoGame 클래스로 캡슐화
